+++ index.html
@@ -113,5 +113,10 @@
       </div>
     </main>
   </body>
+
+  <script type="module" src="keypressOperations.js"></script>
+  <script type="module" src="memoryOperations.js"></script>
+  <script type="module" src="history.js"></script>
+  <script type="module" src="./degreeFunctionExponent.js"></script>
   <script type="module" src="script.js"></script>
 </html>
